LPS-60024 match com.liferay.portal.kernel.template.TemplateContextContributor
